calligra: improve method handling
Before, __getattr__ was used to dynamically instantiate each method upon
attribute access. This caused problems in case of invalid access of unrelated
attribute, where we would look for '__method_'+<attr> and throw a weird exception.
Now, each method is a property and the getter is responsible for dynamically
instantiating the method.